pyproject.toml
[project]
name = "teraflopai-data"
version = "0.1.0"
description = "A petabyte scale data processing framework for AI models using Ray."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Teraflop AI", email = "enrico@teraflop.ai"},
]
classifiers = [
  "Typing :: Typed",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "daft[lance,ray]>=0.5.22",
    "loguru>=0.7.3",
    "ray[data]>=2.47.1",
    "transformers<4.54.0",
]
[project.optional-dependencies]
text = [
    "warcio>=1.7.5",
    "trafilatura>=2.0.0",
    "chonkie>=1.1.0",
    "sentence-transformers>=5.0.0",
    "vllm>=0.9.2",
]
audio = [
    "silero-vad>=5.1.2",
]
video = [
    "av>=15.1.0",
    "scenedetect>=0.6.6",
]
image = [
    "imagehash>=4.3.2",
    "pillow-simd>=9.5.0.post2",
    "vllm>=0.9.2",
]
all = [
    "teraflopai-data[text,audio,video,image]",
]

[tool.uv]
package = true
